pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "daphne>=4.2.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.3",
    "django-prometheus>=2.4.0",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.0",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
    "redis>=6.2.0",
    "uvicorn>=0.34.3",
]
